Move from_binary/1 to test helpers, all tests pass
