+++ test/support/test_helpers.ex
@@ -5,7 +5,7 @@ defmodule Sentry.TestHelpers do
 
   @spec decode_event_from_envelope!(binary()) :: Sentry.Event.t()
   def decode_event_from_envelope!(body) when is_binary(body) do
-    {:ok, %Envelope{items: items}} = Envelope.from_binary(body)
+    {:ok, %Envelope{items: items}} = from_binary(body)
     Enum.find(items, &is_struct(&1, Sentry.Event))
   end
 
@@ -48,4 +48,83 @@ defmodule Sentry.TestHelpers do
   def all_config do
     Enum.sort(for {{:sentry_config, key}, value} <- :persistent_term.get(), do: {key, value})
   end
+
+  def from_binary(binary) when is_binary(binary) do
+    json_library = Sentry.Config.json_library()
+
+    [raw_headers | raw_items] = String.split(binary, "\n")
+
+    with {:ok, headers} <- json_library.decode(raw_headers),
+         {:ok, items} <- decode_items(raw_items, json_library) do
+      {:ok,
+       %Envelope{
+         event_id: headers["event_id"] || nil,
+         items: items
+       }}
+    else
+      {:error, _json_error} -> {:error, :invalid_envelope}
+    end
+  end
+
+  #
+  # Decoding
+  #
+
+  # Steps over the item pairs in the envelope body. The item header is decoded
+  # first so it can be used to decode the item following it.
+  defp decode_items(raw_items, json_library) do
+    items =
+      raw_items
+      |> Enum.chunk_every(2, 2, :discard)
+      |> Enum.map(fn [k, v] ->
+        with {:ok, item_header} <- json_library.decode(k),
+             {:ok, item} <- decode_item(item_header, v, json_library) do
+          item
+        else
+          {:error, _reason} = error -> throw(error)
+        end
+      end)
+
+    {:ok, items}
+  catch
+    {:error, reason} -> {:error, reason}
+  end
+
+  defp decode_item(%{"type" => "event"}, data, json_library) do
+    result = json_library.decode(data)
+
+    case result do
+      {:ok, fields} ->
+        {:ok,
+         %Sentry.Event{
+           breadcrumbs: fields["breadcrumbs"],
+           culprit: fields["culprit"],
+           environment: fields["environment"],
+           event_id: fields["event_id"],
+           source: fields["event_source"],
+           exception: List.wrap(fields["exception"]),
+           extra: fields["extra"],
+           fingerprint: fields["fingerprint"],
+           level: fields["level"],
+           message: fields["message"],
+           modules: fields["modules"],
+           original_exception: fields["original_exception"],
+           platform: fields["platform"],
+           release: fields["release"],
+           request: fields["request"],
+           server_name: fields["server_name"],
+           tags: fields["tags"],
+           timestamp: fields["timestamp"],
+           user: fields["user"]
+         }}
+
+      {:error, e} ->
+        {:error, "Failed to decode event item: #{e}"}
+    end
+  end
+
+  defp decode_item(%{"type" => type}, _data, _json_library),
+    do: {:error, "unexpected item type '#{type}'"}
+
+  defp decode_item(_, _data, _json_library), do: {:error, "Missing item type header"}
 end
